feat(fulfillment): add manual Fulfillment tracking, remove EasyPost and shipping-event dead code
Extracts shipping/fulfillment tracking out of the flat orders.shipping_*
columns into commerce.fulfillments and commerce.shipping_provider_accounts,
each Order having at most one active Fulfillment. Every Store gets a
'manual' shipping provider account at creation; MCP tools
(create_fulfillment, mark_fulfillment_shipped/delivered, cancel_fulfillment,
list_shipping_provider_accounts) drive it end to end through
awaiting_pickup -> shipped -> delivered. orders.shipping_status is now a
read projection recomputed from the Fulfillment rows.

Removes the EasyPost config/env plumbing and the orphaned shipping_events
outbox queue and worker, which never had a real producer and wrote to
columns this change deletes.

Also fixes integration.schedule_analytics_deliveries binding (i64 -> i32
to match the SQL INTEGER parameter).

+++ crates/chaos-core/src/adapters/postgres/analytics/mod.rs
@@ -370,7 +370,7 @@ impl PostgresAnalyticsDeliveryStore {
     pub(crate) async fn schedule_deliveries(&self, limit: u16) -> Result<usize, ApplicationError> {
         let scheduled: Option<i64> =
             sqlx::query_scalar("SELECT integration.schedule_analytics_deliveries($1)")
-                .bind(i64::from(limit))
+                .bind(i32::from(limit))
                 .fetch_one(&self.pool)
                 .await
                 .map_err(db)?;
